move run_cli() up to test/common_util.py
