+++ src/allmydata/test/cli/common.py
@@ -1,9 +1,6 @@
-from cStringIO import StringIO
-from twisted.internet import threads # CLI tests use deferToThread
-from ...util.assertutil import precondition
 from ...util.encodingutil import unicode_to_argv
 from ...scripts import runner
-from ..common_util import ReallyEqualMixin
+from ..common_util import ReallyEqualMixin, run_cli
 
 def parse_options(basedir, command, args):
     o = runner.Options()
@@ -20,18 +17,3 @@ class CLITestMixin(ReallyEqualMixin):
         client_dir = unicode_to_argv(self.get_clientdir(i=client_num))
         nodeargs = [ "--node-directory", client_dir ]
         return run_cli(verb, nodeargs=nodeargs, *args, **kwargs)
-
-def run_cli(verb, *args, **kwargs):
-    precondition(not [True for arg in args if not isinstance(arg, str)],
-                 "arguments to do_cli must be strs -- convert using unicode_to_argv", args=args)
-    nodeargs = kwargs.get("nodeargs", [])
-    argv = nodeargs + [verb] + list(args)
-    stdin = kwargs.get("stdin", "")
-    stdout, stderr = StringIO(), StringIO()
-    d = threads.deferToThread(runner.runner, argv, run_by_human=False,
-                              stdin=StringIO(stdin),
-                              stdout=stdout, stderr=stderr)
-    def _done(rc):
-        return rc, stdout.getvalue(), stderr.getvalue()
-    d.addCallback(_done)
-    return d
